refactor: components/mealFrame・mealFramePattern を meal/frame 階層へ移動、features/mealFrame 削除
- features/mealFrame/ を削除（features/meal/frame/ への移行済み、消し忘れ）
- components/mealFrame/ → components/meal/frame/ へ移動
- components/mealFramePattern/ → components/meal/frame/pattern/ へ移動
- app/ 側の import パスを新しい階層に合わせて更新

これにより components/ と features/ の階層構造が揃い、
バックエンドの Meal::Frame::* 階層とも対応する。
