+++ frontend/src/components/meal/frame/MealFrameList.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from 'react';
 import Link from 'next/link';
-import useMealFrame from '../../features/meal/frame/useMealFrame';
+import useMealFrame from '../../../features/meal/frame/useMealFrame';
 
 const MealFrameList = () => {
   const { mealFrames, deleteMealFrame, refetchMealFrames } = useMealFrame();
